Fixed multiple redundant activity log loads when tab is clicked. Fixes #559
This was because it was binding the event handler to `document` within the `turbolinks:load` event handler.

Since the `document` element persists whilst navigating with turbolinks, the event handler was being bound to it every time the user navigated to a new page.

Any event handlers bound to `document` should only be evaluated once.