Fuzzy matching on event types and eligibilities
This will make it easier to scrape events from different sources, with
eligibility and event type information, without running into validation
issues.

Use a trigram similarity to the dictionary to perform fuzzy matching.
Use the dictionary key and the content of a (array) match field to match
on.

example:

first_come_first_served:
  title: First Come First Served
  match:
    - time
    - first
    - registration order