i18n: allow configuring fallbacks from tess.yml
This commit introduces a simpler way to configure I18n's fallbacks[1]
than Rails' provides, by adding a simple translation layer driven by
two `tess.yml` options. The driver for this is the need for easily
configuring region-specific locales in TeSS, while at the same time
providing for a future source of clean source of desired locales to
show in a UI chooser for users.

Though I18n has a configuration option for available locales (and
Rails too), locales not listed in that option won't be considered for
sources of translations.

[1] Fallback is a concept from I18n that allows to use translations
    from a more general locale than specified, e.g.: a request for a
    translation to locale "en-CA", if not found, will fallback to
    search the translations to "en".

    This enables using a region-specific locale to get easily
    localized dates, but still use the default translations files. It
    also allows to have region-specific with partial coverage
    (i.e.: customizations of some translations only).